PAY-65: Add Model\Member dir + Move LinksTrait to this new folder + Adjust LinksTraitTest to the new namespace
