+++ src/Model/Currency.php
@@ -11,7 +11,7 @@ namespace PayNL\Sdk\Model;
  */
 class Currency implements ModelInterface
 {
-    use LinksTrait;
+    use Member\LinksTrait;
 
     /**
      * @var string
